pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
